refactor: change imports for RoutingComponent
