+++ monolith/service/src/main/kotlin/com/makentoshe/shuvi/service/device/DeleteDeviceService.kt
@@ -1,4 +1,6 @@
-package com.makentoshe.shuvi.service
+package com.makentoshe.shuvi.service.device
+
+import com.makentoshe.shuvi.service.Service
 
 interface DeleteDeviceService : Service {
     override val routing: String
